Made custom table containers for inquiries and users
