+++ frontend/.eslintrc.cjs
@@ -26,5 +26,6 @@ module.exports = {
     'react/jsx-props-no-spreading': 0,
     'import/prefer-default-export': 'warn',
     '@tanstack/query/exhaustive-deps': 0,
+    'import/no-cycle': 0,
   },
 };
