+++ frontend/src/components/index.ts
@@ -12,3 +12,5 @@ export * from './tablecells/DateCell';
 export * from './tablecells/StatusCell';
 export * from './tablecells/RolesCell';
 export * from './tabs/CustomTabPanel';
+export * from './table/InquiryTableContainer';
+export * from './table/UserTableContainer';
